ai_assistant/requirements.txt
openai>=1.0.0
python-dotenv
tenacity
pyttsx3
SpeechRecognition
pyaudio
rich
pyautogui
Pillow
ddgs
python-pptx
